+++ app/routes/components/ScrollArea/index.tsx
@@ -2,8 +2,8 @@ import { ComponentTemplate } from "src/components/ComponentTemplate";
 import { Section } from "src/components/Section";
 import type { Route } from "./+types/index";
 import { changelog } from "./changelog.codegen";
-import { Demo } from "./Demo";
-import { demoSourceCode } from "./Demo_source_code.codegen";
+import Demo01 from "./demos/Demo01";
+import demo01SourceCode from "./demos/Demo01.source_code.codegen";
 import { sourceCode } from "./source_code.codegen";
 
 export function meta(_: Route.MetaArgs) {
@@ -16,12 +16,11 @@ export function meta(_: Route.MetaArgs) {
 export default function ScrollAreaPage() {
   return (
     <ComponentTemplate component="ScrollArea">
-      <Section
-        title="ScrollArea"
-        demoAndCode={[<Demo key="demo" />, demoSourceCode]}
-        sourceCode={sourceCode}
-        changelog={changelog}
-      />
+      <Section.Root title="ScrollArea">
+        <Section.Demo node={<Demo01 />} code={demo01SourceCode} />
+        <Section.SourceCode files={sourceCode} />
+        <Section.Changelog changelog={changelog} />
+      </Section.Root>
     </ComponentTemplate>
   );
 }