+++ src/components/ScrollArea/CHANGELOG.md
@@ -1,2 +1,4 @@
 - 2025-10-04: Initial version.
 - 2025-10-05: Make the scrollbar appear on hover only.
+- 2025-10-10: Add a dragging state for the scrollbar thumb to prevent it from
+  hiding while being dragged.